+++ tests/test_integration.py
@@ -43,6 +43,9 @@ def test_main_argument_passing__all_exists(mock_gemini_qna, caplog, tmp_path) ->
     path_readme.touch()
     os.environ['INPUT_README-PATH'] = str(path_readme)
 
+    # mock return value
+    mock_gemini_qna.return_value = (0, "This is the feedback message.") 
+
     entrypoint.main()
 
     mock_gemini_qna.assert_called_once_with(